genAi/skillshare-api/requirements.txt
fastapi
uvicorn
sqlmodel
passlib
bcrypt==3.2.2
PyJWT
python-dotenv
psycopg2-binary